+++ noodles-gff/src/directive/name.rs
@@ -1,21 +1,24 @@
-mod other;
+//! GFF directive name.
+
+pub mod other;
 mod standard;
 
-pub const GFF_VERSION: Name = Name::Standard(Standard::GffVersion);
-pub const SEQUENCE_REGION: Name = Name::Standard(Standard::SequenceRegion);
-pub const FEATURE_ONTOLOGY: Name = Name::Standard(Standard::FeatureOntology);
-pub const ATTRIBUTE_ONTOLOGY: Name = Name::Standard(Standard::AttributeOntology);
-pub const SOURCE_ONTOLOGY: Name = Name::Standard(Standard::SourceOntology);
-pub const SPECIES: Name = Name::Standard(Standard::Species);
-pub const GENOME_BUILD: Name = Name::Standard(Standard::GenomeBuild);
-pub const FORWARD_REFERENCES_ARE_RESOLVED: Name =
-    Name::Standard(Standard::ForwardReferencesAreResolved);
-pub const START_OF_FASTA: Name = Name::Standard(Standard::StartOfFasta);
+pub use self::other::Other;
+use self::standard::Standard;
 
-use self::{other::Other, standard::Standard};
+pub(super) const GFF_VERSION: Name = Name::Standard(Standard::GffVersion);
+pub(super) const SEQUENCE_REGION: Name = Name::Standard(Standard::SequenceRegion);
+pub(super) const FEATURE_ONTOLOGY: Name = Name::Standard(Standard::FeatureOntology);
+pub(super) const ATTRIBUTE_ONTOLOGY: Name = Name::Standard(Standard::AttributeOntology);
+pub(super) const SOURCE_ONTOLOGY: Name = Name::Standard(Standard::SourceOntology);
+pub(super) const SPECIES: Name = Name::Standard(Standard::Species);
+pub(super) const GENOME_BUILD: Name = Name::Standard(Standard::GenomeBuild);
+pub(super) const FORWARD_REFERENCES_ARE_RESOLVED: Name =
+    Name::Standard(Standard::ForwardReferencesAreResolved);
+pub(super) const START_OF_FASTA: Name = Name::Standard(Standard::StartOfFasta);
 
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub enum Name {
+pub(super) enum Name {
     Standard(Standard),
     Other(Other),
 }